create session_types.protocol package and move ProtocolSession in there. Start protocol executor.
