+++ robot-server/robot_server/service/session/session_types/__init__.py
@@ -4,3 +4,4 @@ from .tip_length_calibration import TipLengthCalibration  # noqa: W0611
 from .deck_calibration_session import DeckCalibrationSession  # noqa: W0611
 from .base_session import BaseSession, SessionMetaData  # noqa: W0611
 from .default_session import DefaultSession  # noqa: W0611
+from .protcol.session import ProtocolSession  # noqa: W0611
